--- src/App.css.backup
@@ -1,269 +1,0 @@
-.app {
-  min-height: 100vh;
-  background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-  position: relative;
-  touch-action: pan-y; /* 세로 스크롤만 허용, 가로 스와이프는 JavaScript에서 처리 */
-  user-select: none; /* 텍스트 선택 방지 */
-  -webkit-user-select: none;
-  -moz-user-select: none;
-  -ms-user-select: none;
-  overflow: hidden; /* 스와이프 시 스크롤바 방지 */
-}
-
-/* Top Header */
-.top-header {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  height: 60px;
-  background: rgba(255, 255, 255, 0.9);
-  backdrop-filter: blur(15px);
-  -webkit-backdrop-filter: blur(15px);
-  border-bottom: 1px solid rgba(255, 255, 255, 0.2);
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 0 20px;
-  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-  z-index: 1000;
-}
-
-.menu-button, .settings-button {
-  background: none;
-  border: none;
-  font-size: 1.5rem;
-  cursor: pointer;
-  padding: 8px;
-  border-radius: 8px;
-  transition: all 0.3s ease;
-  color: #8370FE;
-}
-
-.menu-button:hover, .settings-button:hover {
-  background: rgba(131, 112, 254, 0.1);
-  transform: scale(1.1);
-}
-
-.app-title {
-  font-size: 1.5rem;
-  font-weight: 700;
-  color: #8370FE;
-  margin: 0;
-  cursor: pointer;
-  transition: all 0.3s ease;
-}
-
-.app-title:hover {
-  transform: scale(1.05);
-  text-shadow: 0 2px 8px rgba(131, 112, 254, 0.3);
-}
-
-/* Header Right */
-.header-right {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-}
-
-.user-name {
-  color: #666;
-  font-size: 0.9rem;
-  font-weight: 500;
-}
-
-.logout-button {
-  background: #ff6b6b;
-  color: white;
-  border: none;
-  padding: 6px 12px;
-  border-radius: 6px;
-  font-size: 0.8rem;
-  font-weight: 500;
-  cursor: pointer;
-  transition: all 0.3s ease;
-}
-
-.logout-button:hover {
-  background: #ee5a52;
-  transform: translateY(-1px);
-}
-
-/* Sidebar */
-.sidebar {
-  position: fixed;
-  top: 0;
-  left: 0;
-  height: 100vh;
-  width: 280px;
-  background: rgba(255, 255, 255, 0.95);
-  backdrop-filter: blur(20px);
-  -webkit-backdrop-filter: blur(20px);
-  border-right: 1px solid rgba(255, 255, 255, 0.2);
-  transform: translateX(-100%);
-  transition: transform 0.3s ease;
-  z-index: 1100;
-  box-shadow: 2px 0 20px rgba(0, 0, 0, 0.1);
-}
-
-.sidebar.open {
-  transform: translateX(0);
-}
-
-.sidebar-header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 20px;
-  border-bottom: 2px solid #f8f9fa;
-  background: linear-gradient(135deg, #8370FE 0%, #C0B7FF 100%);
-  color: white;
-}
-
-.sidebar-title {
-  font-size: 1.25rem;
-  font-weight: 600;
-  margin: 0;
-}
-
-.close-button {
-  background: none;
-  border: none;
-  color: white;
-  font-size: 1.5rem;
-  cursor: pointer;
-  padding: 4px;
-  border-radius: 4px;
-  transition: all 0.3s ease;
-}
-
-.close-button:hover {
-  background: rgba(255, 255, 255, 0.2);
-}
-
-.sidebar-nav {
-  padding: 20px 0;
-}
-
-.sidebar-button {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-  width: 100%;
-  padding: 16px 20px;
-  border: none;
-  background: transparent;
-  color: #6c757d;
-  font-size: 1rem;
-  font-weight: 500;
-  cursor: pointer;
-  transition: all 0.3s ease;
-  text-align: left;
-}
-
-.sidebar-button:hover {
-  background: rgba(131, 112, 254, 0.05);
-  color: #8370FE;
-}
-
-.sidebar-button.active {
-  background: linear-gradient(135deg, #8370FE 0%, #C0B7FF 100%);
-  color: white;
-  box-shadow: inset 4px 0 0 rgba(255, 255, 255, 0.3);
-}
-
-.sidebar-icon {
-  font-size: 1.2rem;
-  width: 24px;
-  text-align: center;
-}
-
-.sidebar-label {
-  font-weight: 600;
-}
-
-/* Sidebar Overlay */
-.sidebar-overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background: rgba(0, 0, 0, 0.5);
-  z-index: 1050;
-}
-
-.app-container {
-  max-width: 1400px;
-  margin: 0 auto;
-  padding: 80px 20px 20px 20px; /* Top padding for fixed header */
-  transition: transform 0.3s ease; /* 스와이프 애니메이션 */
-}
-
-/* Main Content */
-.main-content {
-  background: rgba(255, 255, 255, 0.9);
-  backdrop-filter: blur(15px);
-  -webkit-backdrop-filter: blur(15px);
-  border-radius: 20px;
-  padding: 32px;
-  box-shadow: 0 8px 32px rgba(0, 0, 0, 0.1);
-  border: 1px solid rgba(255, 255, 255, 0.2);
-  min-height: 70vh;
-  user-select: text; /* 메인 콘텐츠에서는 텍스트 선택 허용 */
-  -webkit-user-select: text;
-  -moz-user-select: text;
-  -ms-user-select: text;
-}
-
-/* Responsive Design */
-@media (max-width: 768px) {
-  .app-container {
-    padding: 80px 12px 12px 12px;
-  }
-
-  .top-header {
-    padding: 0 12px;
-  }
-
-  .main-content {
-    padding: 20px;
-    border-radius: 16px;
-  }
-
-  .sidebar {
-    width: 100%;
-    max-width: 320px;
-  }
-}
-
-@media (max-width: 480px) {
-  .app-container {
-    padding: 80px 8px 8px 8px;
-  }
-
-  .top-header {
-    padding: 0 8px;
-  }
-
-  .app-title {
-    font-size: 1.25rem;
-  }
-
-  .main-content {
-    padding: 16px;
-    border-radius: 12px;
-  }
-
-  .sidebar {
-    width: 100%;
-  }
-
-  .sidebar-header {
-    padding: 16px;
-  }
-
-  .sidebar-button {
-    padding: 12px 16px;
-  }
-}
