--- src/features/analysis/index.js
@@ -1,5 +1,0 @@
-/**
- * Analysis feature 모듈의 메인 export 파일
- */
-
-export * from './pages';
