--- src/features/analysis/pages/index.js
@@ -1,5 +1,0 @@
-/**
- * Analysis pages 모듈의 메인 export 파일
- */
-
-export { default as AnalysisPage } from './AnalysisPage';
