--- src/features/browse/components/index.js
@@ -1,7 +1,0 @@
-/**
- * Browse components 모듈의 메인 export 파일
- */
-
-export { default as ExplorerPanel, SeriesPanel, CharacterPanel, EpisodePanel } from './ExplorerPanel';
-export { default as Breadcrumb } from './Breadcrumb';
-export { default as DetailView } from './DetailView';
