--- src/features/browse/hooks/index.js
@@ -1,6 +1,0 @@
-/**
- * Browse hooks 모듈의 메인 export 파일
- */
-
-export { useExplorer } from './useExplorer';
-export { useSeriesData, useSeriesCharacters, useSeriesEpisodes } from './useSeriesData';
