--- src/features/browse/index.js
@@ -1,7 +1,0 @@
-/**
- * Browse feature 모듈의 메인 export 파일
- */
-
-export * from './hooks';
-export * from './components';
-export * from './pages';
