--- src/features/browse/pages/index.js
@@ -1,5 +1,0 @@
-/**
- * Browse pages 모듈의 메인 export 파일
- */
-
-export { default as BrowsePage } from './BrowsePage';
