--- src/features/settings/components/index.js
@@ -1,5 +1,0 @@
-/**
- * Settings components 모듈의 메인 export 파일
- */
-
-export { SettingSection, ToggleSwitch, NumberInput } from './SettingSection';
