--- src/features/settings/index.js
@@ -1,6 +1,0 @@
-/**
- * Settings feature 모듈의 메인 export 파일
- */
-
-export * from './components';
-export * from './pages';
