--- src/features/settings/pages/index.js
@@ -1,5 +1,0 @@
-/**
- * Settings pages 모듈의 메인 export 파일
- */
-
-export { default as SettingsPage } from './SettingsPage';
